+++ src/view/slider.rs
@@ -134,7 +134,7 @@ impl View for Slider {
         let x_height = font.x_heights.1 as i32;
 
         let x_drift = if self.value > (self.min_value + self.max_value) / 2.0 {
-            -(small_radius + plan.width as i32)
+            -(small_radius + plan.width)
         } else {
             small_radius
         };
